requirements.txt
beautifulsoup4
faiss-cpu
gradio
langchain_community
langchain_openai
langchainhub
